refactor: Entity 클래스 패키지 이동 (dao -> domain)
