+++ src/main/java/cart/dao/ProductDaoImpl.java
@@ -3,6 +3,8 @@ package cart.dao;
 import java.sql.PreparedStatement;
 import java.util.List;
 import java.util.Optional;
+
+import cart.domain.ProductEntity;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
