+++ src/test/java/cart/dao/ProductDaoImplTest.java
@@ -1,5 +1,6 @@
 package cart.dao;
 
+import cart.domain.ProductEntity;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
